Move the code to events.
